Remove casting warning in parser/parser.h

+++ parser/parser.h
@@ -145,7 +145,7 @@ u8 ba_PAtom(struct ba_Controller* ctr) {
 	// lit_char
 	else if (ba_PAccept(BA_TK_LITCHAR, ctr)) {
 		struct ba_Type type = { BA_TYPE_U8, 0 };
-		ba_PTkStkPush(ctr->pTkStk, (void*)lexVal[0], type, BA_TK_LITCHAR,
+		ba_PTkStkPush(ctr->pTkStk, (void*)(u64)lexVal[0], type, BA_TK_LITCHAR,
 			/* isLValue = */ 0);
 	}
 	// identifier
